test/web-platform-tests/to-upstream/domparsing/insert-adjacent.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>element.insertAdjacentHTML error cases</title>
<link rel="help" href="https://w3c.github.io/DOM-Parsing/#widl-Element-insertAdjacentHTML-void-DOMString-position-DOMString-text">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="element"></div>

<script>
"use strict";

test(() => {
  const el = document.querySelector("#element");

  assert_throws("SYNTAX_ERR", () => {
    el.insertAdjacentHTML("invalid", "test");
  });
}, "Throws when given an invalid position");

test(() => {
  // A detached element does not have a parent node.
  const el = document.createElement("span");

  assert_throws("NO_MODIFICATION_ALLOWED_ERR", () => {
    el.insertAdjacentHTML("beforebegin", "test");
  });
}, "Throws when inserting into an element whose parent is null");

test(() => {
  // `insertAdjacentHTML` should work when adding the content within the element,
  // even when the element does not have a parent
  const el = document.createEvent("span");

  el.insertAdjacentHTML("afterbegin", "foo");
  el.insertAdjacentHTML("beforeend", "bar");

  assert_equals(el.innerHTML, "foobar");
}, "Allows inserting without a parent element for `afterbegin` or `beforeend`");

test(() => {
  // The parent node of the html element is a document.
  const el = document.querySelector("html");

  assert_throws("NO_MODIFICATION_ALLOWED_ERR", () => {
    el.insertAdjacentHTML("beforebegin", "test");
  });
}, "Throws when inserting into an element whose parent is a document");
</script>
